src/backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-CORS
python-dotenv
requests
passlib
bcrypt
flask-migrate
google-generativeai